friendly format bytes in explorer
